refactor
https://coveord.atlassian.net/browse/KIT-1131
