+++ packages/headless/src/utils/expression-builder/common/part.ts
@@ -1,0 +1,3 @@
+export interface Part {
+  toString(): string;
+}
